pull video file to the server no more dynamic polling
Now the user-provided file is uploaded to the server.
It is infeasible to operate dynamically as we convert
file to constant frame rate. Frame timestamps are needed
in advance for CFR export, thus fast-scanning the video.
Ffprobe on user-side could help but then either a custom
video-reader would be needed or we could not provide
seek functionality.